% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{sim_dm}
\alias{sim_dm}
\title{Dail-Madsen Simulator}
\usage{
sim_dm(nsites = 15, nyears = 30, lambda = 100, phi = 0.6, r = 0.4, p = 0.8)
}
\arguments{
\item{nsites}{Integer number of sites. Defaults to 15.}

\item{nyears}{: number of years in the time series}

\item{lambda}{: mean density during t=1}

\item{phi}{: apparent survival (survival)}

\item{r}{: recruitment rate}

\item{p}{: detection probability}
}
\description{
Appendix S4R Code to create population dataFunction to create population data with constant vitalrates (i.e.  C data)
}
\examples{
sim_dm()
}
\keyword{abundance}
